Backend Code Refactoring
To accomodate backend testing, I needed to refactor my code to make it
more readable and understandable. Should have been done earlier, but
nonetheless, it's done.
